+++ apps/vmq_server/src/vmq_ranch_config.erl
@@ -27,7 +27,8 @@
     delete_listener/2,
     restart_listener/2,
     get_listener_config/2,
-    listeners/0
+    listeners/0,
+    listeners/1
 ]).
 
 %% gen_server callbacks
@@ -162,8 +163,17 @@ start_listener(Type, Addr, Port, {SocketOpts, Opts}) ->
                 Error -> Error
             end
     end.
-
 listeners() ->
+    listeners(select, false, false).
+
+listeners(with_tls) ->
+    listeners(select, true, false);
+listeners(with_mqtt) ->
+    listeners(select, false, true);
+listeners(with_tls_and_mqtt) ->
+    listeners(select, true, true).
+
+listeners(select, AddTLSInfo, AddMQTTInfo) ->
     maps:fold(
         fun({Ip, Port}, ConfigMap, Acc) ->
             {ok, {Type, Opts}} = get_listener_config(Ip, Port),
@@ -183,9 +193,55 @@ listeners() ->
                     _ -> inet:ntoa(Ip)
                 end,
             StrPort = integer_to_list(Port),
+            AllowedProtocolVersionsList = proplists:get_value(allowed_protocol_versions, Opts, ""),
+            AllowedProtocolVersions = vmq_util:mqtt_version_to_string(AllowedProtocolVersionsList),
+            AllowAnonymousOverride = proplists:get_value(allow_anonymous_override, Opts, ""),
+            % TLS
+            CertFile = proplists:get_value(certfile, Opts, ""),
+            TLS = proplists:get_value(tls_version, Opts, ""),
+            CAFile = proplists:get_value(ca_file, Opts, ""),
+            KeyFile = proplists:get_value(keyfile, Opts, ""),
+            RequireCertificate = proplists:get_value(require_certificate, Opts, ""),
+            UseIDAsUsername = proplists:get_value(use_identity_as_username, Opts, ""),
+            PSKSupport = proplists:get_value(psk_support, Opts, ""),
+            PSKFile = proplists:get_value(pskfile, Opts, ""),
+
+            Base = [
+                Type,
+                StrIp,
+                StrPort,
+                Status,
+                MountPoint,
+                MaxConnections,
+                ActiveConnections,
+                AllConnections
+            ],
+            Base2 =
+                case AddTLSInfo of
+                    true ->
+                        Base ++
+                            [
+                                TLS,
+                                CertFile,
+                                CAFile,
+                                KeyFile,
+                                RequireCertificate,
+                                UseIDAsUsername,
+                                PSKSupport,
+                                PSKFile
+                            ];
+                    _ ->
+                        Base
+                end,
+            Base3 =
+                case AddMQTTInfo of
+                    true ->
+                        Base2 ++ [AllowedProtocolVersions, AllowAnonymousOverride];
+                    _ ->
+                        Base2
+                end,
             [
-                {Type, StrIp, StrPort, Status, MountPoint, MaxConnections, ActiveConnections,
-                    AllConnections}
+                list_to_tuple(Base3)
                 | Acc
             ]
         end,